backend/pyproject.toml
[project]
name = "mysite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "faker>=40.4.0",
    "gunicorn>=25.0.3",
    "pillow>=12.1.1",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "ipython>=9.9.0",
]
